make header compile
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@4662 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ fem/space/common/adaptcallbackhandle.hh
@@ -42,7 +42,7 @@ namespace Dune
       rpOp_( org.rpOp_ )
     {}
 
-    void preAdapt ( const size_t estimatedAdditionalElements )
+    void preAdapt ( const unsigned int estimatedAdditionalElements )
     {
       dofManager_.reserveMemory( estimatedAdditionalElements );
     }
